requirements.txt
psycopg2==2.4.1
Django==1.3.1
PIL==1.1.7
South==0.7.3
google-api-python-client==1.0beta4
httplib2==0.7.1
oauth2client==1.0beta4
python-gflags==1.6
simplejson==2.2.1
sorl-thumbnail==11.09
wsgiref==0.1.2

python-memcached
ipython